requirements.txt
-i https://pypi.org/simple
--extra-index-url https://bloomberg.bintray.com/pip/simple
numpy>=1.15.0
pandas>=0.23.4
pdblp>=0.1.7
pyarrow>=0.11.1
pytz>=2018.7
pytest>=4.0.1
PyYAML>=3.13
